requirements.txt
# Core OCR dependencies for high-performance text extraction
# PaddlePaddle framework (CPU version for compatibility)
paddlepaddle>=2.5.0

# OCR engine with state-of-the-art performance
paddleocr>=2.7.0

# ONNX Runtime for inference acceleration (CPU version)
onnxruntime>=1.15.0

# Computer vision library for image preprocessing
opencv-python-headless>=4.8.0

# NumPy for array operations and image processing
numpy>=1.24.0

# Optional: Pillow for additional image format support
Pillow>=9.5.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0